First shot at exec_bg
Add a new function exec_bg() that can take a command to execute
and ensure that any processes forked inside the exec does not
hang the php process.

This is done by ensuring output is redirected to a file and then
read back. That way PHP will not wait around for background processes
to finish.

Fixes: https://github.com/elkarbackup/elkarbackup/issues/646