Fix client/job/config route
https://github.com/elkarbackup/elkarbackup/issues/597 documents an
error in the `showJobConfig()` function.

This seems related to a symfony update which restricts access to
the full container.

`container->get('kernel')` does not work anymore and needs to be handled
by injecting the kernel into the current object.
Luckily this is already done, so all we have to do is use the existing
`$this->kernel` object.