Fix spurious test failure for withInvalidNotificationLevel
The test was checking if the API correctly refuses invalid notification levels.
For this, the test was randomly generating levels with `rand(0,199)`.
Unfortunately, 0 is a valid level so if the random generator returns 0, the check
would pass even though a fail is expected.

Fix this to generate random numbers from 1 to 199 inclusively, removing this 0.51%
chance of hitting this test failure.