Try to run php unit tests in a better fashion
The current testsuite runs phpunit from bin, which is the stub loader
installed by composer.

When this is run, composer will install the requirements for phpunit and
error with a message about not finding a composer.json.
This is because the composer.json is in /app/elkarbackup and not in
presumably ./bin/.

Try to fix this by forcing phpUnit to be installed as part of the dev
environment and use the binary in ./vendor/

Drive-By: Pull in some php dependency updates.